+++ lib/compiler/src/expression/function_call.rs
@@ -309,7 +309,7 @@ impl<'a> Builder<'a> {
                                 // The variable kind is expected to be equal to
                                 // the ind of the target of the closure.
                                 VariableKind::Target => {
-                                    (target.type_info(state).result, target.as_value())
+                                    (target.type_info(state).result, target.resolve_constant())
                                 }
 
                                 // The variable kind is expected to be equal to